norme ok but leaks to fix ToT

+++ linux/srcs/win_utils.c
@@ -15,7 +15,8 @@
 int	quit_program(t_env *env)
 {
 	write(1, "Goodbye!", 7);
-	destroy_all_images(env);
+	if (&(env->floor) != NULL)
+		destroy_all_images(env);
 	mlx_destroy_image(env->params.mlx, env->img.img);
 	mlx_destroy_window(env->params.mlx, env->params.mlx_win);
 	mlx_destroy_display(env->params.mlx);
@@ -23,38 +24,40 @@ int	quit_program(t_env *env)
 	exit(1);
 }
 
+int	move_1(int key, t_env *env)
+{
+	if (key == UP || key == ARROW_UP)
+		move_player(env, &(env->player_back), 0, -1);
+	if (key == DOWN || key == ARROW_DOWN)
+		move_player(env, &(env->player_front), 0, 1);
+	if (key == LEFT || key == ARROW_LEFT)
+		move_player(env, &(env->player_left), -1, 0);
+	if (key == RIGHT || key == ARROW_RIGHT)
+		move_player(env, &(env->player_right), 1, 0);
+	return (0);
+}
+
+int	move_2(int key, t_env *env)
+{
+	if (key == UP || key == ARROW_UP)
+		move_player(env, &(env->player_back2), 0, -1);
+	if (key == DOWN || key == ARROW_DOWN)
+		move_player(env, &(env->player_front2), 0, 1);
+	if (key == LEFT || key == ARROW_LEFT)
+		move_player(env, &(env->player_left2), -1, 0);
+	if (key == RIGHT || key == ARROW_RIGHT)
+		move_player(env, &(env->player_right2), 1, 0);
+	return (0);
+}
+
 int	keypress(int key, t_env *env)
 {
 	if (key == ESC)
 		quit_program(env);
-	if (key == UP || key == ARROW_UP)
-	{	
-		if (env->count % 2 == 0)
-			move_player(env, &(env->player_back), 0, -1);
-		else
-			move_player(env, &(env->player_back2), 0, -1);
-	}
-	if (key == DOWN || key == ARROW_DOWN)
-	{
-		if (env->count % 2 == 0)
-			move_player(env, &(env->player_front), 0, 1);
-		else
-			move_player(env, &(env->player_front2), 0, 1);
-	}
-	if (key == LEFT || key == ARROW_LEFT)
-	{
-		if (env->count % 2 == 0)
-			move_player(env, &(env->player_left), -1, 0);
-		else
-			move_player(env, &(env->player_left2), -1, 0);
-	}
-	if (key == RIGHT || key == ARROW_RIGHT)
-	{
-		if (env->count % 2 == 0)
-			move_player(env, &(env->player_right), 1, 0);
-		else
-			move_player(env, &(env->player_right2), 1, 0);
-	}
+	if (env->count % 2 == 0)
+		move_1(key, env);
+	else
+		move_2(key, env);
 	env->count++;
 	printf("%i\n", env->count);
 	return (0);
